requirements.txt
python-telegram-bot>=20.7,<22.0
anthropic>=0.40.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
